dirty windows build support
