+++ cmd/rdpgw/protocol/gateway_windows.go
